Separates unit and e2e test suites. Prepares for feature based test suites.
