+++ smsapi/mms_e2e_test.go
@@ -1,7 +1,8 @@
-package e2e
+// +build e2e
+
+package smsapi
 
 import (
-	"github.com/smsapi/smsapi-go/smsapi"
 	"log"
 	"testing"
 	"time"
@@ -25,7 +26,7 @@ func TestSendMms(t *testing.T) {
 }
 
 func TestScheduleMms(t *testing.T) {
-	future := &smsapi.Timestamp{time.Now().Local().AddDate(0, 0, 2)}
+	future := &Timestamp{time.Now().Local().AddDate(0, 0, 2)}
 
 	ctx, cancel := createCtx()
 	defer cancel()
