+++ smsapi/sendername_e2e_test.go
@@ -1,4 +1,6 @@
-package e2e
+// +build e2e
+
+package smsapi
 
 import (
 	"log"
