+++ smsapi/vms_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package smsapi
 
 import (
